rep walk: 64-bit byte-decomposition perm LUT on host AND device (N <= 64)
The canonicalization walk (dim x terms x |G| permutation images per matvec)
is THE production hot loop, and both its implementations were stranded on
the serial n_sites-iteration bit walk in exactly the regime that matters:
the host byte-LUT fast path was uint32-valued and hard-capped at N <= 32,
and the device policy never had a LUT at all -- at the 36-site 126M-dim
block that is 36 dependent loads per image x 72 images x 432 terms x 1.26e8
rows = 26 s per GPU matvec, ~10.6 h per star.

* RepSectorData::perm_lut_data -> uint64 words, built for every N <= 64
  (5 byte-chunks at N=36; ~740 KB at |G|=72 -- L2-resident).
* Host RepSymmetryBasisPolicy::apply_perm consumes the 64-bit table.
* DeviceRepSymmetryBasisPolicy gains the same fast path (5 unrolled
  gathers replace the 36-iteration loop); the mirror builder reuses the
  caller's host-built table or builds one from perms_flat.

Validated: 4x4 labeled rows bit-identical to the dense-path reference;
5x5 Lanczos-path levels identical (and 15% faster on CPU despite 25 sites
already having had the 32-bit path). GPU timing measurement deferred to
the hardware-validation job.

+++ include/ed/matvec/rep_symmetry_basis_policy.h
@@ -81,9 +81,10 @@ struct RepSymmetryBasisPolicy {
 
     // Optional byte-decomposition LUT for fast apply_perm (N≤32).
     // When non-null, replaces the N-iteration scalar bit-scatter loop with
-    // ``perm_lut_bpw`` table lookups (~4 for N=32). Pointer into
-    // RepSectorData::perm_lut_data; null when N>32 or not built.
-    const std::uint32_t* perm_lut     = nullptr;
+    // ``perm_lut_bpw`` table lookups (5 for N=36). Pointer into
+    // RepSectorData::perm_lut_data; null when not built. 64-bit words so
+    // every N <= 64 rides the fast path.
+    const std::uint64_t* perm_lut     = nullptr;
     int                  perm_lut_bpw = 0;
 
     [[nodiscard]] inline std::uint64_t dim() const noexcept { return dim_; }
@@ -106,12 +107,12 @@ struct RepSymmetryBasisPolicy {
     apply_perm(std::uint64_t s, int g) const noexcept {
         const std::uint64_t flip = (flips != nullptr) ? flips[g] : 0ULL;
         if (perm_lut != nullptr) {
-            const std::uint32_t* lut_g = perm_lut
+            const std::uint64_t* lut_g = perm_lut
                 + static_cast<std::size_t>(g) * perm_lut_bpw * 256;
-            std::uint32_t r = 0;
+            std::uint64_t r = 0;
             for (int b = 0; b < perm_lut_bpw; ++b)
                 r |= lut_g[b * 256 + static_cast<int>((s >> (b * 8)) & 0xFF)];
-            return static_cast<std::uint64_t>(r) ^ flip;
+            return r ^ flip;
         }
         // Scalar fallback for N>32 (or when LUT not built).
         const int* p = perms + static_cast<std::size_t>(g) * n_sites;
